app ai/requirements.txt
streamlit
requests
beautifulsoup4
tldextract
python-whois
matplotlib
pandas
